backend/requirements.txt
Flask==3.1.1
psycopg2-binary
Jinja2
Werkzeug==3.1.0
